split multiple object types into separate files rename graph file restructure a bit
