+++ src/domain/community/virtual-contributor/dto/prompt-graph-definition/prompt.graph.definition.data.struct.dto.ts
@@ -1,0 +1,14 @@
+import { ObjectType, Field } from '@nestjs/graphql';
+import { PromptGraphDefinitionDataPoint } from './prompt.graph.definition.data.point.dto';
+
+@ObjectType()
+export class PromptGraphDefinitionDataStruct {
+  @Field({ nullable: true })
+  title?: string;
+
+  @Field({ nullable: true })
+  type?: string;
+
+  @Field(() => [PromptGraphDefinitionDataPoint], { nullable: true })
+  properties?: PromptGraphDefinitionDataPoint[];
+}
