+++ src/domain/community/virtual-contributor/dto/prompt-graph-definition/prompt.graph.definition.edge.dto.ts
@@ -1,0 +1,10 @@
+import { ObjectType, Field } from '@nestjs/graphql';
+
+@ObjectType()
+export class PromptGraphDefinitionEdge {
+  @Field({ nullable: true })
+  from?: string;
+
+  @Field({ nullable: true })
+  to?: string;
+}
