+++ src/services/ai-server/ai-persona/ai.persona.service.ts
@@ -22,7 +22,7 @@ import { EncryptionService } from '@hedger/nestjs-encryption';
 import { AiPersonaEngineAdapterInvocationInput } from '../ai-persona-engine-adapter/dto/ai.persona.engine.adapter.dto.invocation.input';
 import { IAiServer } from '../ai-server/ai.server.interface';
 import { AiPersonaEngine } from '@common/enums/ai.persona.engine';
-import graphJson from './dto/expert.graph.json';
+import graphJson from '../prompt-graph/config/prompt.graph.expert.json';
 
 @Injectable()
 export class AiPersonaService {
