+++ src/services/ai-server/prompt-graph/dto/prompt.graph.data.point.dto.ts
@@ -1,0 +1,16 @@
+import { ObjectType, Field } from '@nestjs/graphql';
+
+@ObjectType()
+export class PromptGraphDataPoint {
+  @Field(() => String, { nullable: false })
+  name!: string;
+
+  @Field(() => String, { nullable: true })
+  type?: string;
+
+  @Field({ nullable: true })
+  description?: string;
+
+  @Field(() => Boolean, { nullable: true })
+  optional?: boolean;
+}
