+++ src/services/ai-server/prompt-graph/dto/prompt.graph.edge.dto.ts
@@ -1,0 +1,10 @@
+import { ObjectType, Field } from '@nestjs/graphql';
+
+@ObjectType()
+export class PromptGraphEdge {
+  @Field({ nullable: true })
+  from?: string;
+
+  @Field({ nullable: true })
+  to?: string;
+}
